+++ trace-processor/applications/sink/trace-ingestor/src/main/java/eu/csgroup/coprs/monitoring/traceingestor/TraceIngestorProperties.java
@@ -4,7 +4,10 @@ import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @Data
-@ConfigurationProperties(prefix = "trace-ingestor")
+@ConfigurationProperties(prefix = "ingestion")
 public class TraceIngestorProperties {
-
+    /**
+     * Configuration file path containing strategy to ingest trace. Must be an absolute path prefixed with 'file:'
+     */
+    private String path;
 }