+++ trace-processor/common/src/main/java/eu/csgroup/coprs/monitoring/common/datamodel/entities/ClonableEntity.java
@@ -2,4 +2,6 @@ package eu.csgroup.coprs.monitoring.common.datamodel.entities;
 
 public interface ClonableEntity<T> {
     T copy();
+
+    void setId(Long id);
 }