+++ trace-processor/common/src/main/java/eu/csgroup/coprs/monitoring/common/datamodel/entities/DefaultEntity.java
@@ -1,0 +1,4 @@
+package eu.csgroup.coprs.monitoring.common.datamodel.entities;
+
+public abstract class DefaultEntity implements ClonableEntity {
+}