+++ trace-processor/common/src/main/java/eu/csgroup/coprs/monitoring/common/datamodel/entities/ExternalInput.java
@@ -1,6 +1,6 @@
 package eu.csgroup.coprs.monitoring.common.datamodel.entities;
 
-import com.vladmihalcea.hibernate.type.basic.PostgreSQLEnumType;
+import com.vladmihalcea.hibernate.type.json.JsonBinaryType;
 import eu.csgroup.coprs.monitoring.common.bean.AutoMergeableMap;
 import lombok.Data;
 import org.hibernate.annotations.*;
@@ -11,17 +11,15 @@ import java.time.Instant;
 
 @Data
 @Entity
+@TypeDef(name = "jsonb", typeClass = JsonBinaryType.class)
 @Inheritance(strategy = InheritanceType.JOINED)
-@TypeDefs({
-    @TypeDef(name = "pgsql_enum", typeClass = PostgreSQLEnumType.class),
-})
-public abstract class ExternalInput implements ClonableEntity {
+public abstract class ExternalInput extends DefaultEntity {
     @Id
     @SequenceGenerator(sequenceName="external_input_id_seq", name = "external_input_id_seq", allocationSize=1)
     @GeneratedValue(generator = "external_input_id_seq", strategy = GenerationType.SEQUENCE)
     private Long id;
 
-    //@Column(nullable = false, unique = true)
+    @Column(nullable = false, unique = true)
     private String filename;
 
     private String mission;
@@ -34,7 +32,7 @@ public abstract class ExternalInput implements ClonableEntity {
 
     private Instant catalogStorageDate;
 
-    @Type( type = "json" )
-    @Column(columnDefinition = "json")
+    @Type( type = "jsonb" )
+    @Column(columnDefinition = "jsonb")
     private AutoMergeableMap custom;
 }