+++ trace-processor/common/src/main/java/eu/csgroup/coprs/monitoring/common/jpa/ProductRepository.java
@@ -1,0 +1,6 @@
+package eu.csgroup.coprs.monitoring.common.jpa;
+
+import eu.csgroup.coprs.monitoring.common.datamodel.entities.Product;
+
+public interface ProductRepository extends EntityRepository<Product> {
+}